Feat: 주문이 있을 경우 상품 수정, 삭제 불가능

+++ src/main/java/com/chaewsscode/chaewsstore/product/service/ProductService.java
@@ -1,12 +1,15 @@
 package com.chaewsscode.chaewsstore.product.service;
 
 import com.chaewsscode.chaewsstore.domain.Account;
+import com.chaewsscode.chaewsstore.domain.CustomerOrder;
 import com.chaewsscode.chaewsstore.domain.Product;
+import com.chaewsscode.chaewsstore.exception.BadRequestException;
 import com.chaewsscode.chaewsstore.exception.ForbiddenException;
 import com.chaewsscode.chaewsstore.exception.ResourceNotFoundException;
 import com.chaewsscode.chaewsstore.product.controller.dto.ProductResponseDto;
 import com.chaewsscode.chaewsstore.product.controller.dto.UpdateProductRequestDto;
 import com.chaewsscode.chaewsstore.product.service.dto.ProductServiceDto;
+import com.chaewsscode.chaewsstore.repository.CustomerOrderRepository;
 import com.chaewsscode.chaewsstore.repository.ProductRepository;
 import com.chaewsscode.chaewsstore.util.ResponseCode;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final CustomerOrderRepository orderRepository;
 
     @Transactional(readOnly = true)
     public Page<ProductResponseDto> readProducts(Pageable pageable) {
@@ -49,6 +53,10 @@ public class ProductService {
         if (!product.getAccount().equals(account)) {
             throw new ForbiddenException(ResponseCode.UPDATE_PRODUCT_FAIL_NOT_OWNER);
         }
+        // 해당 상품의 주문 내역이 있는지 확인
+        if (orderRepository.existsByProduct(product)) {
+            throw new BadRequestException(ResponseCode.UPDATE_PRODUCT_FAIL_ALREADY_SOLDOUT);
+        }
         product.setInfo(request.getPrice(), request.getIsSold());
 
         return ProductResponseDto.of(product);
@@ -62,6 +70,10 @@ public class ProductService {
         // 상품 주인 확인
         if (!product.getAccount().equals(account)) {
             throw new ForbiddenException(ResponseCode.DELETE_PRODUCT_FAIL_NOT_OWNER);
+        }
+        // 해당 상품의 주문 내역이 있는지 확인
+        if (orderRepository.existsByProduct(product)) {
+            throw new BadRequestException(ResponseCode.DELETE_PRODUCT_FAIL_ALREADY_SOLDOUT);
         }
         productRepository.delete(product);
     }
